+++ util/src/browser.ts
@@ -1,3 +1,5 @@
-import * as browser from './_browser';
-export * from './_browser';
-export { browser as default };
+export * from '~/shared/index';
+export * from '~/client/index';
+
+export * from './browser';
+export * as default from './browser';