+++ util/src/index.ts
@@ -1,3 +1,6 @@
-import * as project from './_project';
-export * from './_project';
-export { project as default };
+export * from '~/shared/index';
+export * from '~/client/index';
+export * from '~/server/index';
+
+export * from './index';
+export * as default from './index';