+++ util/src/server.ts
@@ -1,3 +1,0 @@
-import * as server from './_server';
-export * from './_server';
-export { server as default };